requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
langchain-community
langchain-core
langchain-groq